# cmd/server/docs/swagger.yaml
basePath: /
definitions:
...
  dto.EquipmentItem:
    properties:
      artifact:
        type: boolean
      attack:
        type: integer
      createdAt:
        type: string
      defense:
        type: integer
      hp:
        type: integer
      id:
        type: string
      image:
        type: string
      name:
        type: string
      price:
        type: integer
      requiredLevel:
        type: integer
      slug:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Game API for Moonshine
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Moonshine API
  version: "1.0"
paths:
...
  /api/equipment_items:
    get:
      consumes:
      - application/json
      description: Get list of equipment items by category
      parameters:
      - description: Equipment category
        in: query
        name: category
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.EquipmentItem'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get equipment items
      tags:
      - equipment
  /api/equipment_items/{slug}/buy:
    post:
      consumes:
      - application/json
      description: Purchase an equipment item
      parameters:
      - description: Item slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Buy equipment item
      tags:
      - equipment
  /api/equipment_items/{slug}/sell:
    post:
      consumes:
      - application/json
      description: Sell an equipment item from inventory
      parameters:
      - description: Item slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Sell equipment item
      tags:
      - equipment
  /api/equipment_items/{slug}/take_on:
    post:
      consumes:
      - application/json
      description: Equip an equipment item from inventory
      parameters:
      - description: Item slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Equip an item
      tags:
      - equipment
  /api/equipment_items/take_off/{slot}:
    post:
      consumes:
      - application/json
      description: Remove an equipment item from equipped slot
      parameters:
      - description: Equipment slot
        in: path
        name: slot
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Unequip an item
      tags:
      - equipment
  /api/fights/current:
    get:
      consumes:
      - application/json
      description: Get information about current active fight
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get current fight
      tags:
      - fights
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: 'JWT token. Example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"